extract some components and minor visual fixes
